improvements from TODOs completed
