+++ examples/secure_cloud_function_internal_server/outputs.tf
